Api Intergration for the frontend

+++ main-app/src/server/middleware/auth.ts
@@ -1,11 +1,12 @@
 import { PrismaClient } from '@prisma/client';
 import {
-  createError,
   defineEventHandler,
-  deleteCookie,
-  getSession,
   H3Event,
   EventHandlerRequest,
+  getSession,
+  handleCors,
+  createError,
+  deleteCookie,
 } from 'h3';
 
 const prisma = new PrismaClient();
@@ -17,57 +18,50 @@ const { SESSION_COOKIE_NAME, SESSION_SECRET } = import.meta.env;
  */
 
 export default defineEventHandler(async (event): Promise<object | void> => {
-  const { id: session_token, data } = await getSession(event, {
+  handleCors(event, {
+    credentials: true,
+  });
+
+  const session = await getSession(event, {
     password: SESSION_SECRET,
     name: SESSION_COOKIE_NAME,
+    cookie: {
+      maxAge: 60 * 60 * 24 * import.meta.env.SESSION_EXPIRY_DAYS,
+    },
   });
-  const { user_id, _ip } = data;
-  if (session_token) {
-    if (!user_id) {
-      event.context['session'] = null;
-    } else {
-      await manageSession(event, session_token, user_id);
-    }
-  } else {
-    throw createError({
-      statusCode: 401,
-      statusMessage: 'Unauthorized',
-    });
+
+  const { id: session_token } = session;
+  if (!session_token) {
+    event.context['session'] = null;
+    return;
   }
+
+  await manageSession(event, session_token);
 });
 
 const manageSession = async (
   event: H3Event<EventHandlerRequest>,
   session_token: string,
-  user_id: string,
 ) => {
   const session = await prisma.session.findUnique({
     where: {
       session_token,
-      user_id,
     },
   });
-  if (session) {
-    if (session.expires.toISOString() < new Date().toISOString()) {
-      await prisma.session.delete({
-        where: {
-          id: session.id,
-        },
-      });
-      deleteCookie(event, SESSION_COOKIE_NAME);
-      if (!event.path.endsWith('/signup') && !event.path.endsWith('/signin')) {
-        throw createError({
-          statusCode: 401,
-          statusMessage: 'Session expired',
-        });
-      }
-    } else {
-      event.context['session'] = session;
-    }
-  } else {
-    throw createError({
-      statusCode: 401,
-      statusMessage: 'Unauthorized',
+
+  if (session && session.expires.toISOString() <= new Date().toISOString()) {
+    deleteCookie(event, SESSION_COOKIE_NAME);
+    await prisma.session.delete({
+      where: {
+        id: session.id,
+      },
     });
+    if (!event.path.endsWith('/signup') && !event.path.endsWith('/signin')) {
+      throw createError({
+        statusCode: 401,
+        statusMessage: 'Session expired',
+      });
+    }
   }
+  event.context['session'] = session;
 };
